aws.load_balancer.py splited and TargetGroup class moved to aws.target_group.py file
